scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-datastore>=2.24.0",
]
